+++ opapi/src/section-types-generator/index.ts
@@ -1,3 +1,3 @@
-export type { DefaultState } from './section-types-generator.types'
-export { executeSectionParsers, executeOperationParsers, composeFilesFromBlocks } from './section-types-generator'
+export type { DefaultState } from './types'
+export { executeSectionParsers, executeOperationParsers, composeFilesFromBlocks } from './generator'
 export { sectionParsers, operationParsers } from './constants'
